requirements/base.txt
Django
django-braces
django-environ
django-model-utils
Pillow
psycopg2
stripe
